some misc fixes/cleanup
git-svn-id: http://svn.macosforge.org/repository/ruby/MacRuby/branches/icu@3632 23306eb0-4c56-4727-a40e-e92c0eb68959

+++ include/ruby/intern.h
@@ -588,7 +588,7 @@ VALUE rb_str_append(VALUE, VALUE);
 VALUE rb_str_concat(VALUE, VALUE);
 VALUE rb_str_plus(VALUE str1, VALUE str2);
 long rb_memhash(const void *ptr, long len);
-int rb_str_hash(VALUE);
+unsigned long rb_str_hash(VALUE);
 int rb_str_hash_cmp(VALUE,VALUE);
 int rb_str_comparable(VALUE, VALUE);
 int rb_str_cmp(VALUE, VALUE);
